Updated ActiveRecord::Acts::Money to not break dynamic finders when not supplied a currency.
The failure was due to the mappings that were being passed into #composed_of. For each mapping, a database column is expected to be found by ActiveRecord when it tries to match on dynamic finders. This removes the idea of a default currency on the ActiveRecord side of things. Folks can use Money.default_currency= to change things globally. If they want to provide a custom currency they can always supply the database field and supply the :currency option.